Moving redis messengers to messenger crate
