+++ messenger/Cargo.toml
@@ -10,6 +10,13 @@ documentation = "https://docs.rs/solana-geyser-plugin-interface"
 [dependencies]
 solana-geyser-plugin-interface  = { version = "=1.10.10" }
 plerkle-serialization={path= "../plerkle_serialization"}
+redis = {version = "0.21.5", features = ["aio", "tokio-comp", "streams"], optional = true}
+log = "0.4.11"
+thiserror = "1.0.30"
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
+
+[features]
+default = ["redis"]
+redis = ["dep:redis"]
