+++ plerkle/Cargo.toml
@@ -13,7 +13,6 @@ documentation = "https://docs.rs/solana-geyser-plugin-interface"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-redis = {version = "0.21.5", features = ["aio", "tokio-comp", "streams"]}
 log = "0.4.11"
 async-trait = "0.1.53"
 solana-sdk = { version = "=1.10.10" }
