--- plerkle/src/constants.rs
@@ -1,4 +1,0 @@
-// Redis stream values.
-pub const GROUP_NAME: &str = "plerkle";
-pub const CONSUMER_NAME: &str = "ingester";
-pub const DATA_KEY: &str = "data";
